Исправлены ошибки:
1. Неправильный тип данных был (беззнаковый целый)
2. Добавили исключения
3. Добавили перевод из типа enum в целый тип данных

+++ Task 2/input_int.jl
@@ -1,5 +1,5 @@
 function input_int(message = "")
     println(message);
     str = readline();
-    return tryparse(UInt8, str);
+    return tryparse(Int128, str);
 end
